Refactor orchestration into library and add integration tests
- move orchestration flow into a reusable library entry point
- add logging initializer and adjust CLI to use it
- cover end-to-end behavior with new integration tests
- preserve existing CLI behavior while improving testability
